Add support for ctrl and shift in keyboard.luad
Since we have discovered that key+ctrl is key + 512 and key + shift
is key + 256, we can create an embedded table for each key + ctrl or
shift or both. So with this commit

      keyboard.ctrl.a
      keyboard.ctrl.shift.a
      keyboard.shift.a
      keyboard.shift.ctrl.a

become available. Yes, keyboard.ctrl.shift.a and keyboard.shift.ctrl.a
are duplicated but it's not costly to duplicate them and makes easier
for whoever is writing (does not need to remember which one is the right
one)

Change-Id: I45367c9b1e9f803d7e62fdfcd25b84426943a836